mf-recommender/requirements.txt
duckdb==1.2.0
Faker==36.1.1
Flask==3.1.0
numpy==2.0.2
torch==2.6.0
torchvision==0.21.0
pytorch-ignite==0.5.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
scikit-learn==1.6.1
flask-cors==5.0.1